Rename V5 protocol module to V6, since V5 and V6 are exactly the same with respect to the WS messages.
